Working on tests for fat jar.
