+++ mleap-databricks-runtime-testkit/src/main/scala/ml/combust/mleap/databricks/runtime/testkit/TestXgboost.scala
@@ -1,0 +1,9 @@
+package ml.combust.mleap.databricks.runtime.testkit
+
+import org.apache.spark.sql.SparkSession
+
+class TestXgboost(session: SparkSession) extends Runnable {
+  override def run(): Unit = {
+    println("Still need to test")
+  }
+}
